% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/componentAnalysis_pca.R
\name{componentAnalysis.pca}
\alias{componentAnalysis.pca}
\title{componentAnalysis.pca}
\format{
[R6::R6Class] object.
}
\description{
Performs a pca anlysis on a given dataset.
The explained variance ratio is calculated
and used to indicate the optimal number of components for a pca.
}
\details{
Performs a pca anlysis on a given dataset.
The explained variance ratio is calculated
and used to indicate the optimal number of components for a pca.
}
\examples{
library(tidyverse)
library(pguXAI)

main = function(){
  # load data set and remove class labels
  df_data <- iris \%>\%
    dplyr::select(-Species)

  # define true class labels
  classes_true <- iris$Species

  # run principle component analysis
  ca <- pguXAI::componentAnalysis.pca$new()
  ca$train(obj=df_data)

  # plot results
  ca$evr_overview_plot()

  ca$evr_detail_plot() \%>\%
    show()

  ca$cos2_corrplot() \%>\%
    show()

  ca$cos2_barplot()

  ca$contrib_corrplot() \%>\%
    show()

  ca$contrib_barplot()

  fin <- "done"
  fin
}

main()

}
\author{
Sebastian Malkusch, \email{malkusch@med.uni-frankfurt.de}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{rslt_pca}}{Returns the instance variable rslt.pca
(list)}

\item{\code{df_pca}}{Returns the instance variable df_pca
(tibble::tibble)}

\item{\code{nComp}}{Returns the instance varibale nComp
(integer)}

\item{\code{verbose}}{Returns the instance variable verbose
(logical)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{componentAnalysis.pca$new()}}
\item \href{#method-finalize}{\code{componentAnalysis.pca$finalize()}}
\item \href{#method-print}{\code{componentAnalysis.pca$print()}}
\item \href{#method-train}{\code{componentAnalysis.pca$train()}}
\item \href{#method-evr_overview_plot}{\code{componentAnalysis.pca$evr_overview_plot()}}
\item \href{#method-evr_detail_plot}{\code{componentAnalysis.pca$evr_detail_plot()}}
\item \href{#method-cos2_corrplot}{\code{componentAnalysis.pca$cos2_corrplot()}}
\item \href{#method-cos2_barplot}{\code{componentAnalysis.pca$cos2_barplot()}}
\item \href{#method-contrib_corrplot}{\code{componentAnalysis.pca$contrib_corrplot()}}
\item \href{#method-contrib_barplot}{\code{componentAnalysis.pca$contrib_barplot()}}
\item \href{#method-clone}{\code{componentAnalysis.pca$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates and returns a new componentAnalysis.pca object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{componentAnalysis.pca$new(verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{Makes the class chatty.
Default is FALSE.
(logical)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new R6 object of type componentAnalysis.pca.
(pguXAI::componentAnalysis.pca)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
Clears the heap and
indicates that instance of componentAnalysis.pca is removed from heap.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{componentAnalysis.pca$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Prints instance variables of a componentAnalysis.pca object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{componentAnalysis.pca$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-train"></a>}}
\if{latex}{\out{\hypertarget{method-train}{}}}
\subsection{Method \code{train()}}{
trains the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{componentAnalysis.pca$train(obj = "tbl_df")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{The data to be analyzed.
(tibble::tibble)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-evr_overview_plot"></a>}}
\if{latex}{\out{\hypertarget{method-evr_overview_plot}{}}}
\subsection{Method \code{evr_overview_plot()}}{
Plots the explained variance ration of the analysis.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{componentAnalysis.pca$evr_overview_plot()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(ggplot2::ggplot)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-evr_detail_plot"></a>}}
\if{latex}{\out{\hypertarget{method-evr_detail_plot}{}}}
\subsection{Method \code{evr_detail_plot()}}{
Plots the eigenvalues of the analysis.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{componentAnalysis.pca$evr_detail_plot()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(ggplot2::ggplot)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cos2_corrplot"></a>}}
\if{latex}{\out{\hypertarget{method-cos2_corrplot}{}}}
\subsection{Method \code{cos2_corrplot()}}{
Plots the quality of representation of the variables on factor map.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{componentAnalysis.pca$cos2_corrplot()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(ggplot2::ggplot)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cos2_barplot"></a>}}
\if{latex}{\out{\hypertarget{method-cos2_barplot}{}}}
\subsection{Method \code{cos2_barplot()}}{
Plots the quality of representation of the variables on factor map.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{componentAnalysis.pca$cos2_barplot()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(ggplot2::ggplot)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-contrib_corrplot"></a>}}
\if{latex}{\out{\hypertarget{method-contrib_corrplot}{}}}
\subsection{Method \code{contrib_corrplot()}}{
Maps the attributes contribution to each component.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{componentAnalysis.pca$contrib_corrplot()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(ggplot2::ggplot)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-contrib_barplot"></a>}}
\if{latex}{\out{\hypertarget{method-contrib_barplot}{}}}
\subsection{Method \code{contrib_barplot()}}{
Plots the quality of representation of the variables on factor map.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{componentAnalysis.pca$contrib_barplot()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(ggplot2::ggplot)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{componentAnalysis.pca$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
