% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossVal_rpart.R
\name{crossVal.rpart}
\alias{crossVal.rpart}
\title{crossVal.rpart}
\format{
[R6::R6Class] object.
}
\description{
A Cross Validation for the CART tree.
}
\details{
Some mor details will follow
}
\examples{
data("iris")
df_data = tibble::as_tibble(iris)
ctrl <- rpart::rpart.control(minsplit = 3, maxdepth = 3)
cv <- pguXAI::crossVal.rpart$new(obj = df_data, label = "Species", ctrl = ctrl, k = 10)
plot(cv)

}
\author{
Sebastian Malkusch, \email{malkusch@med.uni-frankfurt.de}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{df_data}}{Returns the instance variable df_data
(tibble::tibble)}

\item{\code{label}}{Returns the instance variable label
(character)}

\item{\code{seed}}{Returns the instance variable seed
(integer)}

\item{\code{lvls}}{Returns the instance variable lvls
(character)}

\item{\code{k}}{Returns the instance variable k
(integer)}

\item{\code{ctrl}}{Returns the instance variable ctrl
(rpart::rpart.control)}

\item{\code{auc_roc}}{Returns the instance variable auc_roc
(numeric)}

\item{\code{confusion_matrix}}{Returns the instance variable confucion_matrix}

\item{\code{scores}}{Returns the cross validation scores
(tibble::tibble)}

\item{\code{accuracy}}{Returns the corss validation accuracy
(numeric)}

\item{\code{verbose}}{Returns the corss validation verbose
(logical)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{crossVal.rpart$new()}}
\item \href{#method-train}{\code{crossVal.rpart$train()}}
\item \href{#method-finalize}{\code{crossVal.rpart$finalize()}}
\item \href{#method-print}{\code{crossVal.rpart$print()}}
\item \href{#method-plot}{\code{crossVal.rpart$plot()}}
\item \href{#method-clone}{\code{crossVal.rpart$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates and returns a new crossVal.rpart object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crossVal.rpart$new(
  obj = "tbl_df",
  label = "character",
  ctrl = "rpart::rpart.control",
  k = 10,
  seed = 42,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{The data to be analyzed.
(tibble::tibble)}

\item{\code{label}}{The column name of the labels within obj
(character)}

\item{\code{ctrl}}{The control sequence for the CART tree
(rpart::rpart.control)}

\item{\code{k}}{The corss validation parameter
(integer)}

\item{\code{seed}}{The seed for the cross validation
(integer)}

\item{\code{verbose}}{Makes the class print results
(logical)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new crossVal.rpart object.
(pguXAI::crossVal.rpart)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-train"></a>}}
\if{latex}{\out{\hypertarget{method-train}{}}}
\subsection{Method \code{train()}}{
trains the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crossVal.rpart$train(repeats = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{repeats}}{The repeat parameter
(integer)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
Clears the heap and
indicates that instance of crossVal.rpart is removed from heap.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crossVal.rpart$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Prints instance variables of a crossVal.rpart object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crossVal.rpart$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
Plots the confusion matrix of the cross validation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crossVal.rpart$plot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crossVal.rpart$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
