% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridSearchCV_rpart.R
\name{gridSearchCV.rpart}
\alias{gridSearchCV.rpart}
\title{gridSearchCV.rpart}
\format{
[R6::R6Class] object.
}
\description{
A cross validation grid search for hyperparameters of the CART tree.
}
\details{
Some mor details will follow
}
\examples{
data("iris")
df_data = tibble::as_tibble(iris)
ctrl <- rpart::rpart.control(minsplit = 3, maxdepth = 3)
cv <- pguXAI::crossVal.cart$new(obj = df_data, label = "Species", ctrl = ctrl, k = 10)
plot(cv)

}
\author{
Sebastian Malkusch, \email{malkusch@med.uni-frankfurt.de}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{data_df}}{Returns the instance variable data_df
(tibble::tibble)}

\item{\code{label}}{Returns the instance variable label
(character)}

\item{\code{seed}}{Returns the instance variable seed
(integer)}

\item{\code{k}}{Returns the instance variable k
(integer)}

\item{\code{repeats}}{Returns the instance variable repeats
(integer)}

\item{\code{grid}}{Returns the instance variable grid
(tibble::tibble)}

\item{\code{scores}}{Returns the instance variable scores
(tbl_df)}

\item{\code{verbose}}{Returns the instance variable verbose
(logical)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{gridSearchCV.rpart$new()}}
\item \href{#method-finalize}{\code{gridSearchCV.rpart$finalize()}}
\item \href{#method-print}{\code{gridSearchCV.rpart$print()}}
\item \href{#method-calc_model_scores}{\code{gridSearchCV.rpart$calc_model_scores()}}
\item \href{#method-train}{\code{gridSearchCV.rpart$train()}}
\item \href{#method-best_model_parameter}{\code{gridSearchCV.rpart$best_model_parameter()}}
\item \href{#method-clone}{\code{gridSearchCV.rpart$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates and returns a new gridSearchCV.rpart object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gridSearchCV.rpart$new(
  obj = "tbl_df",
  label = "character",
  grid = "list",
  k = 10,
  repeats = 1,
  seed = 42,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{The data to be analyzed.
(tibble::tibble)}

\item{\code{label}}{The column name of the labels within obj
(character)}

\item{\code{grid}}{A list comprising the hyperparameter variations.
(list)}

\item{\code{k}}{The corss validation parameter
(integer)}

\item{\code{repeats}}{The repeat parameter
(integer)}

\item{\code{seed}}{The seed for the cross validation
(integer)}

\item{\code{verbose}}{Makes the class print results
(logical)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new gridSearchCV.rpart object.
(pguXAI::gridSearchCV.rpart)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
Clears the heap and
indicates that instance of gridSearchCV.rpart is removed from heap.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gridSearchCV.rpart$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Prints instance variables of a gridSearchCV.rpart object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gridSearchCV.rpart$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calc_model_scores"></a>}}
\if{latex}{\out{\hypertarget{method-calc_model_scores}{}}}
\subsection{Method \code{calc_model_scores()}}{
creates a crossVal.rpart model and returns its scores
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gridSearchCV.rpart$calc_model_scores(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{The ctrl parameter of the model.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-train"></a>}}
\if{latex}{\out{\hypertarget{method-train}{}}}
\subsection{Method \code{train()}}{
runs the grid search
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gridSearchCV.rpart$train()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-best_model_parameter"></a>}}
\if{latex}{\out{\hypertarget{method-best_model_parameter}{}}}
\subsection{Method \code{best_model_parameter()}}{
Returns the parameters of the best model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gridSearchCV.rpart$best_model_parameter(score = "character")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{score}}{The score parameter by which the optimum is defined
(character)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gridSearchCV.rpart$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
