% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaKMeans.R
\name{pca.KMeans}
\alias{pca.KMeans}
\title{pca.KMeans}
\format{
[R6::R6Class] object.
}
\description{
Performs KMeans clustering on a given pca dataset.
}
\details{
Performs KMeans clustering on a given pca dataset.
}
\examples{

data_df <- iris \%>\%
   dplyr::select(-Species)

nComponents <- 2
PreProcessor <- caret::preProcess(x=df_data, method=c("center", "scale", "pca"), pcaComp = nComponents)
df_pred <- predict(PreProcessor, df_data)

km <- pguXAI::pca.KMeans$new(n=3, seed = 42, verbose = TRUE)
km$train(obj = df_pred, n = 100)

km$probHist_plot() \%>\%
  plot()
km$cluster_plot(obj = df_pred)
km$df_centers \%>\%
  print()

}
\author{
Sebastian Malkusch, \email{malkusch@med.uni-frankfurt.de}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{seed}}{Returns the instance variable seed
(integer)}

\item{\code{setSeed}}{Sets the instance variable seed
(integer)}

\item{\code{nCenters}}{Returns the instance variable nCenters
(integer)}

\item{\code{setNCenters}}{Sets the instance variable nCenters
(integer)}

\item{\code{level}}{Returns the instancs variable level
(character)}

\item{\code{predProb}}{Returns th instance variable predProb
(matrix)}

\item{\code{predClass}}{Returns the instance variable predClass
(factor)}

\item{\code{df_centers}}{Returns the instance variable df_centers
(tbl_df)}

\item{\code{verbose}}{Returns the instance variable verbose
(logical)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{pca.KMeans$new()}}
\item \href{#method-finalize}{\code{pca.KMeans$finalize()}}
\item \href{#method-print}{\code{pca.KMeans$print()}}
\item \href{#method-train}{\code{pca.KMeans$train()}}
\item \href{#method-cluster_kmeans}{\code{pca.KMeans$cluster_kmeans()}}
\item \href{#method-cluster_statistics}{\code{pca.KMeans$cluster_statistics()}}
\item \href{#method-probHist_plot}{\code{pca.KMeans$probHist_plot()}}
\item \href{#method-cluster_plot}{\code{pca.KMeans$cluster_plot()}}
\item \href{#method-silhouette_plot}{\code{pca.KMeans$silhouette_plot()}}
\item \href{#method-clone}{\code{pca.KMeans$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates and returns a new pca.KMeans object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pca.KMeans$new(n = 2, seed = 42, verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Initial number of cluster
(integer)}

\item{\code{seed}}{An initial seed.
Default is 42
(integer)}

\item{\code{verbose}}{Makes the class chatty.
Default is FALSE.
(logical)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new R6 object of type pca.KMeans.
(pguXAI::pca.KMeans)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
Clears the heap and
indicates that instance of pca.KMeans is removed from heap.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pca.KMeans$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Prints instance variables of a pca.KMeans object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pca.KMeans$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-train"></a>}}
\if{latex}{\out{\hypertarget{method-train}{}}}
\subsection{Method \code{train()}}{
trains the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pca.KMeans$train(obj = "tbl_df", n = 100)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{The data to be analyzed.
Needs to be the result of a pca analysis.
(tibble::tibble)}

\item{\code{n}}{Indicates the number of training iterations
(integer)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cluster_kmeans"></a>}}
\if{latex}{\out{\hypertarget{method-cluster_kmeans}{}}}
\subsection{Method \code{cluster_kmeans()}}{
Performs iterative kMeans step.
Not to run by the user.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pca.KMeans$cluster_kmeans(
  obj = "tbl_df",
  classes_init = "factor",
  n = "integer"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{The data to be analyzed.
Needs to be the result of a pca analysis.
(tibble::tibble)}

\item{\code{classes_init}}{Contains the true class labels of the data in obj
(factor)}

\item{\code{n}}{Indicates the number of training iterations
(integer)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cluster_statistics"></a>}}
\if{latex}{\out{\hypertarget{method-cluster_statistics}{}}}
\subsection{Method \code{cluster_statistics()}}{
Performs iterative kMeans step.
Not to run by the user.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pca.KMeans$cluster_statistics(obj = "tbl_df")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{The data to be analyzed.
Needs to be the result of a pca analysis.
(tibble::tibble)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-probHist_plot"></a>}}
\if{latex}{\out{\hypertarget{method-probHist_plot}{}}}
\subsection{Method \code{probHist_plot()}}{
Plots a histogram of Class probability.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pca.KMeans$probHist_plot()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(list)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cluster_plot"></a>}}
\if{latex}{\out{\hypertarget{method-cluster_plot}{}}}
\subsection{Method \code{cluster_plot()}}{
Plots Clustering Result in all pca dimensions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pca.KMeans$cluster_plot(obj = "tbl_df")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{The data to be analyzed.
Needs to be the result of a pca analysis.
(tibble::tibble)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(list)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-silhouette_plot"></a>}}
\if{latex}{\out{\hypertarget{method-silhouette_plot}{}}}
\subsection{Method \code{silhouette_plot()}}{
Plots Silhouette analysis
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pca.KMeans$silhouette_plot(obj = "tbl_df")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{The data to be analyzed.
Needs to be the result of a pca analysis.
(tibble::tibble)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(list)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pca.KMeans$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
