% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenKMeans.R
\name{screen.KMeans}
\alias{screen.KMeans}
\title{pca.KMeans}
\format{
[R6::R6Class] object.
}
\description{
Performs KMeans clustering on a given pca dataset.
}
\details{
Performs KMeans clustering on a given pca dataset.
}
\examples{

}
\author{
Sebastian Malkusch, \email{malkusch@med.uni-frankfurt.de}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{seed}}{Returns the instance variable seed
(integer)}

\item{\code{setSeed}}{Sets the instance variable seed
(integer)}

\item{\code{maxCluster}}{Returns the instance variable maxCluster
(integer)}

\item{\code{setMaxCluster}}{Sets the instance variable maxCluster
(integer)}

\item{\code{df_data}}{Returns the instance variable df_data
(tbl_df)}

\item{\code{df_wcss}}{Returns the instance variable df_wcss
(tbl_df)}

\item{\code{df_silhouette}}{Returns the instance variable df_silhouette
(tbl_df)}

\item{\code{verbose}}{Returns the instance variable verbose
(logical)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{screen.KMeans$new()}}
\item \href{#method-finalize}{\code{screen.KMeans$finalize()}}
\item \href{#method-print}{\code{screen.KMeans$print()}}
\item \href{#method-train}{\code{screen.KMeans$train()}}
\item \href{#method-kmeansAnalysis}{\code{screen.KMeans$kmeansAnalysis()}}
\item \href{#method-wcssAnalysis}{\code{screen.KMeans$wcssAnalysis()}}
\item \href{#method-silhouetteAnalysis}{\code{screen.KMeans$silhouetteAnalysis()}}
\item \href{#method-wcss_boxplot}{\code{screen.KMeans$wcss_boxplot()}}
\item \href{#method-wcss_mean_plot}{\code{screen.KMeans$wcss_mean_plot()}}
\item \href{#method-wcss_deriv1_plot}{\code{screen.KMeans$wcss_deriv1_plot()}}
\item \href{#method-wcss_deriv2_plot}{\code{screen.KMeans$wcss_deriv2_plot()}}
\item \href{#method-wcss_component_plot}{\code{screen.KMeans$wcss_component_plot()}}
\item \href{#method-silhouette_boxplot}{\code{screen.KMeans$silhouette_boxplot()}}
\item \href{#method-silhouette_mean_plot}{\code{screen.KMeans$silhouette_mean_plot()}}
\item \href{#method-silhouette_deriv1_plot}{\code{screen.KMeans$silhouette_deriv1_plot()}}
\item \href{#method-silhouette_deriv2_plot}{\code{screen.KMeans$silhouette_deriv2_plot()}}
\item \href{#method-silhouette_component_plot}{\code{screen.KMeans$silhouette_component_plot()}}
\item \href{#method-clone}{\code{screen.KMeans$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates and returns a new screen.KMeans object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{screen.KMeans$new(n = 2, seed = 42, verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Initial number of maxCluster
(integer)}

\item{\code{seed}}{An initial seed.
Default is 42
(integer)}

\item{\code{verbose}}{Makes the class chatty.
Default is FALSE.
(logical)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new R6 object of type pca.KMeans.
(pguXAI::pca.KMeans)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
Clears the heap and
indicates that instance of pca.KMeans is removed from heap.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{screen.KMeans$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Prints instance variables of screen.KMeans object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{screen.KMeans$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-train"></a>}}
\if{latex}{\out{\hypertarget{method-train}{}}}
\subsection{Method \code{train()}}{
trains the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{screen.KMeans$train(obj = "tbl_df", n = 10)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{The data to be analyzed.
Needs to be the result of a pca analysis.
(tibble::tibble)}

\item{\code{n}}{number of iterations
(integer)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-kmeansAnalysis"></a>}}
\if{latex}{\out{\hypertarget{method-kmeansAnalysis}{}}}
\subsection{Method \code{kmeansAnalysis()}}{
Performs an iterative KMeans step.
Not to run by the user.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{screen.KMeans$kmeansAnalysis(obj, n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{The data to be analyzed.
Needs to be the result of a pca analysis.
(tibble::tibble)}

\item{\code{n}}{Seed additive
(integer)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Result dataframe of the kmeans analysis.
(tbl_df)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wcssAnalysis"></a>}}
\if{latex}{\out{\hypertarget{method-wcssAnalysis}{}}}
\subsection{Method \code{wcssAnalysis()}}{
Performs wcss analysis on df_data.
Not to run by the user.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{screen.KMeans$wcssAnalysis()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-silhouetteAnalysis"></a>}}
\if{latex}{\out{\hypertarget{method-silhouetteAnalysis}{}}}
\subsection{Method \code{silhouetteAnalysis()}}{
Performs silhouette analysis on df_data.
Not to run by the user.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{screen.KMeans$silhouetteAnalysis()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wcss_boxplot"></a>}}
\if{latex}{\out{\hypertarget{method-wcss_boxplot}{}}}
\subsection{Method \code{wcss_boxplot()}}{
Creates a boxplot of the wcss vs the number of clusters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{screen.KMeans$wcss_boxplot()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Boxplot of wcss screening result
(ggplot2::ggplot)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wcss_mean_plot"></a>}}
\if{latex}{\out{\hypertarget{method-wcss_mean_plot}{}}}
\subsection{Method \code{wcss_mean_plot()}}{
Creates a plot of the screening result wcss mean values vs
the number of clusters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{screen.KMeans$wcss_mean_plot()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Plot of WCSS mean.
(ggplot2::ggplot)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wcss_deriv1_plot"></a>}}
\if{latex}{\out{\hypertarget{method-wcss_deriv1_plot}{}}}
\subsection{Method \code{wcss_deriv1_plot()}}{
Creates a plot of the first derivation of the screening result of
the wcss mean values vs the number.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{screen.KMeans$wcss_deriv1_plot()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Plot of first derivation of WCSS.
(ggplot2::ggplot)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wcss_deriv2_plot"></a>}}
\if{latex}{\out{\hypertarget{method-wcss_deriv2_plot}{}}}
\subsection{Method \code{wcss_deriv2_plot()}}{
Creates a plot of the second derivation of the screening result of
the wcss mean values vs the number.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{screen.KMeans$wcss_deriv2_plot()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Plot of second derivation of WCSS.
(ggplot2::ggplot)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wcss_component_plot"></a>}}
\if{latex}{\out{\hypertarget{method-wcss_component_plot}{}}}
\subsection{Method \code{wcss_component_plot()}}{
summarizes the wcss results of the srceening analysis in a component plot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{screen.KMeans$wcss_component_plot()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Component plot
(ggplot2::ggplot)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-silhouette_boxplot"></a>}}
\if{latex}{\out{\hypertarget{method-silhouette_boxplot}{}}}
\subsection{Method \code{silhouette_boxplot()}}{
Creates a boxplot of the silhouette scores vs the number of clusters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{screen.KMeans$silhouette_boxplot()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Boxplot of silhouette score screening result
(ggplot2::ggplot)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-silhouette_mean_plot"></a>}}
\if{latex}{\out{\hypertarget{method-silhouette_mean_plot}{}}}
\subsection{Method \code{silhouette_mean_plot()}}{
Creates a plot of the screening result silhouette score mean values vs
the number of clusters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{screen.KMeans$silhouette_mean_plot()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Plot of silhouette score mean.
(ggplot2::ggplot)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-silhouette_deriv1_plot"></a>}}
\if{latex}{\out{\hypertarget{method-silhouette_deriv1_plot}{}}}
\subsection{Method \code{silhouette_deriv1_plot()}}{
Creates a plot of the first derivation of the screening result of
the silhouette score mean values vs the number.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{screen.KMeans$silhouette_deriv1_plot()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Plot of first derivation of silhouette score.
(ggplot2::ggplot)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-silhouette_deriv2_plot"></a>}}
\if{latex}{\out{\hypertarget{method-silhouette_deriv2_plot}{}}}
\subsection{Method \code{silhouette_deriv2_plot()}}{
Creates a plot of the second derivation of the screening result of
the silhouette score mean values vs the number.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{screen.KMeans$silhouette_deriv2_plot()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Plot of second derivation of silhouette score.
(ggplot2::ggplot)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-silhouette_component_plot"></a>}}
\if{latex}{\out{\hypertarget{method-silhouette_component_plot}{}}}
\subsection{Method \code{silhouette_component_plot()}}{
summarizes the silhouette score results of the srceening analysis in a component plot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{screen.KMeans$silhouette_component_plot()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Component plot
(ggplot2::ggplot)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{screen.KMeans$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
