% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/componentAnalysis_pca.R
\name{componentAnalysis.pca}
\alias{componentAnalysis.pca}
\title{componentAnalysis.pca}
\format{
[R6::R6Class] object.
}
\description{
Performs a pca anlysis on a given dataset.
The explained variance ratio is calculated
and used to indicate the optimal number of components for a pca.
}
\details{
Performs a pca anlysis on a given dataset.
The explained variance ratio is calculated
and used to indicate the optimal number of components for a pca.
}
\examples{
data_df <- iris \%>\%
   dplyr::select(-Species)
ca <- pguXAI::componentAnalysis.pca$new(thr = 0.99, verbose=FALSE)
ca$train(obj=data_df)
print(ca)
plot(ca)

}
\author{
Sebastian Malkusch, \email{malkusch@med.uni-frankfurt.de}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{threshold}}{Returns the instance variable threshold
(numeric)}

\item{\code{setThreshold}}{Sets the instance variable threshold
(numeric)}

\item{\code{df_pca}}{Returns the instance variable df_pca
(tibble::tibble)}

\item{\code{nComp}}{Returns the instance varibale nComp
(integer)}

\item{\code{verbose}}{Returns the instance variable verbose
(logical)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{componentAnalysis.pca$new()}}
\item \href{#method-finalize}{\code{componentAnalysis.pca$finalize()}}
\item \href{#method-print}{\code{componentAnalysis.pca$print()}}
\item \href{#method-train}{\code{componentAnalysis.pca$train()}}
\item \href{#method-sd_plot}{\code{componentAnalysis.pca$sd_plot()}}
\item \href{#method-evr_plot}{\code{componentAnalysis.pca$evr_plot()}}
\item \href{#method-cumSumEvr_plot}{\code{componentAnalysis.pca$cumSumEvr_plot()}}
\item \href{#method-plot}{\code{componentAnalysis.pca$plot()}}
\item \href{#method-clone}{\code{componentAnalysis.pca$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates and returns a new componentAnalysis.pca object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{componentAnalysis.pca$new(thr = 0.95, verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{thr}}{An initial threshold.
Default is 0.95
(numeric)}

\item{\code{verbose}}{Makes the class chatty.
Default is FALSE.
(logical)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new R6 object of type componentAnalysis.pca.
(pguXAI::componentAnalysis.pca)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
Clears the heap and
indicates that instance of componentAnalysis.pca is removed from heap.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{componentAnalysis.pca$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Prints instance variables of a componentAnalysis.pca object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{componentAnalysis.pca$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-train"></a>}}
\if{latex}{\out{\hypertarget{method-train}{}}}
\subsection{Method \code{train()}}{
trains the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{componentAnalysis.pca$train(obj = "tbl_df")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{The data to be analyzed.
(tibble::tibble)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sd_plot"></a>}}
\if{latex}{\out{\hypertarget{method-sd_plot}{}}}
\subsection{Method \code{sd_plot()}}{
Plots standard deviation of detected compenents.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{componentAnalysis.pca$sd_plot()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(list)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-evr_plot"></a>}}
\if{latex}{\out{\hypertarget{method-evr_plot}{}}}
\subsection{Method \code{evr_plot()}}{
Plots the explained variance ratio of the components
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{componentAnalysis.pca$evr_plot()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(list)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cumSumEvr_plot"></a>}}
\if{latex}{\out{\hypertarget{method-cumSumEvr_plot}{}}}
\subsection{Method \code{cumSumEvr_plot()}}{
Plots the cummulative sum of the explained variance ratio of the components
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{componentAnalysis.pca$cumSumEvr_plot()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(list)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
Plots the analysis result.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{componentAnalysis.pca$plot()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(list)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{componentAnalysis.pca$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
