% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaKMeans.R
\name{pca.KMeans}
\alias{pca.KMeans}
\title{pca.KMeans}
\format{
[R6::R6Class] object.
}
\description{
Performs KMeans clustering on a given pca dataset.
}
\details{
Performs KMeans clustering on a given pca dataset.
}
\examples{
library(tidyverse)
library(pguXAI)
library(FactoMineR)
library(caret)

main = function(){
  # load data set and remove class labels
  df_data <- iris \%>\%
    dplyr::select(-Species)

  # define true class labels
  classes_true <- iris$Species

  # define nuber of components for pca and number of clusters for kmeans
  nComponents <- 2
  nCluster <- 10

  # pre-scale the data for pca
  PreProcessor <- caret::preProcess(x=df_data, method=c("center", "scale"), pcaComp = nComponents)
  df_scaled <- predict(PreProcessor, df_data)

  # reduce dimensions of sclaed dataset using pca
  rslt_pca <- df_scaled \%>\%
    FactoMineR::PCA(ncp = nComponents, scale.unit = FALSE, graph = FALSE)
  df_pred <- as.data.frame(predict(rslt_pca, df_scaled)$coord)

  # run kmeans analysis
  km <- pguXAI::pca.KMeans$new(n=nCluster, seed = 42, verbose = TRUE)
  km$train(obj = df_pred)

  km$cluster_plot(obj = df_pred)

  km$silhouette_plot(obj = df_pred) \%>\%
    plot()

  print("Result of silhouette analysis:")
  km$df_silhouette \%>\%
    print()

  print("Average silhouette width:")
  km$av_sil_width \%>\%
    print()

  print("Centers of clusters:")
  km$df_centers \%>\%
    print()

  print("Probability of the class label assignment:")
  km$predProb \%>\%
    print()

  print("Majority vote of the class label assignment:")
  km$predClass \%>\%
    print()

  print("Within cluster sum of squares analysis:")
  km$df_withinss \%>\%
    print()

  km$tot_withinss \%>\%
    print()

  fin <- "done"
  fin
}

main()

}
\author{
Sebastian Malkusch, \email{malkusch@med.uni-frankfurt.de}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{seed}}{Returns the instance variable seed
(integer)}

\item{\code{setSeed}}{Sets the instance variable seed
(integer)}

\item{\code{nCenters}}{Returns the instance variable nCenters
(integer)}

\item{\code{setNCenters}}{Sets the instance variable nCenters
(integer)}

\item{\code{level}}{Returns the instancs variable level
(character)}

\item{\code{predClass}}{Returns the instance variable predClass
(factor)}

\item{\code{df_centers}}{Returns the instance variable df_centers
(tbl_df)}

\item{\code{df_silhouette}}{Returns the instance variable df_silhouette
(tbl_df)}

\item{\code{av_sil_width}}{Returns the instance variable av_sil_width
(numeric)}

\item{\code{av_withinss}}{Returns the instance variable av_withinss
(tbl_df)}

\item{\code{tot_withinss}}{Returns the instance variable tot_withinss
(numeric)}

\item{\code{verbose}}{Returns the instance variable verbose
(logical)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{pca.KMeans$new()}}
\item \href{#method-finalize}{\code{pca.KMeans$finalize()}}
\item \href{#method-print}{\code{pca.KMeans$print()}}
\item \href{#method-train}{\code{pca.KMeans$train()}}
\item \href{#method-cluster_statistics}{\code{pca.KMeans$cluster_statistics()}}
\item \href{#method-silhouette_analysis}{\code{pca.KMeans$silhouette_analysis()}}
\item \href{#method-cluster_plot}{\code{pca.KMeans$cluster_plot()}}
\item \href{#method-silhouette_plot}{\code{pca.KMeans$silhouette_plot()}}
\item \href{#method-clone}{\code{pca.KMeans$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates and returns a new pca.KMeans object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pca.KMeans$new(n = 2, seed = 42, verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Initial number of cluster
(integer)}

\item{\code{seed}}{An initial seed.
Default is 42
(integer)}

\item{\code{verbose}}{Makes the class chatty.
Default is FALSE.
(logical)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new R6 object of type pca.KMeans.
(pguXAI::pca.KMeans)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
Clears the heap and
indicates that instance of pca.KMeans is removed from heap.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pca.KMeans$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Prints instance variables of a pca.KMeans object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pca.KMeans$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-train"></a>}}
\if{latex}{\out{\hypertarget{method-train}{}}}
\subsection{Method \code{train()}}{
trains the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pca.KMeans$train(obj = "tbl_df")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{The data to be analyzed.
Needs to be the result of a pca analysis.
(tibble::tibble)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cluster_statistics"></a>}}
\if{latex}{\out{\hypertarget{method-cluster_statistics}{}}}
\subsection{Method \code{cluster_statistics()}}{
Performs cluster analysis step..
Not to run by the user.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pca.KMeans$cluster_statistics(obj = "tbl_df")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{The data to be analyzed.
Needs to be the result of a pca analysis.
(tibble::tibble)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-silhouette_analysis"></a>}}
\if{latex}{\out{\hypertarget{method-silhouette_analysis}{}}}
\subsection{Method \code{silhouette_analysis()}}{
Performs a silouette analysis.
Not to run by the user.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pca.KMeans$silhouette_analysis(obj = "tbl_df")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{The data to be analyzed.
Needs to be the result of a pca analysis.
(tibble::tibble)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cluster_plot"></a>}}
\if{latex}{\out{\hypertarget{method-cluster_plot}{}}}
\subsection{Method \code{cluster_plot()}}{
Plots Clustering Result in all pca dimensions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pca.KMeans$cluster_plot(obj = "tbl_df")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{The data to be analyzed.
Needs to be the result of a pca analysis.
(tibble::tibble)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(list)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-silhouette_plot"></a>}}
\if{latex}{\out{\hypertarget{method-silhouette_plot}{}}}
\subsection{Method \code{silhouette_plot()}}{
Plots Silhouette analysis
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pca.KMeans$silhouette_plot(obj = "tbl_df")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{The data to be analyzed.
Needs to be the result of a pca analysis.
(tibble::tibble)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(list)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pca.KMeans$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
